updates for logging and some bug fixes

+++ src/app/views.py
@@ -2,6 +2,7 @@
 import json
 import traceback
 
+from django.conf import settings
 from django.core.paginator import Paginator
 from django.core.serializers.json import DjangoJSONEncoder
 from django.forms import formset_factory, modelform_factory
@@ -20,6 +21,8 @@ from django.views.decorators.cache import cache_page
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_GET, require_POST
 
+from app.utils.custom_logging import CustomLogging
+
 from .algorithm import algorithms as alg
 from .forms import *
 from .models import Algorithm, Passage, Word
@@ -30,8 +33,10 @@ from .providers.passage_provider import passage_provider
 from .providers.word_provider import word_provider
 from .utils import references
 
+logger = CustomLogging(__name__)
 
-def search(request: HttpRequest) -> HttpResponse:
+
+def search_page(request: HttpRequest) -> HttpResponse:
     context = {}
     # words = word_provider.get_verbs([1437611, 1437621])
     # print(words)
@@ -42,101 +47,126 @@ def search(request: HttpRequest) -> HttpResponse:
 # TODO SEE: https://docs.djangoproject.com/en/4.2/topics/cache/
 # AND https://docs.djangoproject.com/en/4.2/topics/http/sessions/
 # @cache_page(60 * 15)
-def read(request: HttpRequest) -> HttpResponse:
-    params = references.parse_reference(request.GET)
-    # Display Genesis 1 if the reference isn't specified.
-    reference = params if params[0] else [1] * 2 + [None] * 3
+def read_page(request: HttpRequest) -> HttpResponse:
+    context = {}
+    try:
+        params = references.parse_reference(request.GET)
+        # Display Genesis 1 if the reference isn't specified.
+        reference = params if params[0] else [1] * 2 + [None] * 3
 
-    words_loaded = word_provider.load_words_if_not_added()
-    words = word_provider.get_words_from_reference(reference)
-    books = book_provider.get_all_book_instances(as_json=True)
-    reference_string = references.get_reference_string(words, abbreviation=True)
+        words_loaded = word_provider.load_words_if_not_added()
+        words = word_provider.get_words_from_reference(reference)
+        books = book_provider.get_all_book_instances(as_json=True)
+        reference_string = references.get_reference_string(words, abbreviation=True)
 
-    context = {
-        "words": word_provider.words_to_json(words),
-        "reference": reference_string,
-        "books": books,
-        "words_loaded": words_loaded,
-    }
+        context = {
+            "words": word_provider.words_to_json(words),
+            "reference": reference_string,
+            "books": books,
+            "words_loaded": words_loaded,
+        }
+        logger.info(f"Read page used: {reference_string}", request)
+    except Exception as e:
+        logger.error(f"Read page error {e}", request, exc_info=True)
     return render(request, "read.html", context)
 
 
 # @cache_page(60 * 15)
-def passages(request: HttpRequest) -> HttpResponse:
-    book = request.GET.get("book")
-    # book_index = BOOK_TO_INDEX.get(book)
+def passages_page(request: HttpRequest) -> HttpResponse:
+    context = {}
+    try:
+        book = request.GET.get("book")
+        # book_index = BOOK_TO_INDEX.get(book)
 
-    passages_loaded = passage_provider.load_passages_if_not_added()
-    passages = passage_provider.get_all_passages(as_json=True)
-    books = book_provider.get_all_book_instances(as_json=True)
-    # paginator = Paginator(passages, 10)  # Show 10 objects per page
-    # page_number = request.GET.get("page")
-    # page_obj = paginator.get_page(page_number)
+        passages_loaded = passage_provider.load_passages_if_not_added()
+        passages = passage_provider.get_all_passages(as_json=True)
+        books = book_provider.get_all_book_instances(as_json=True)
+        # paginator = Paginator(passages, 10)  # Show 10 objects per page
+        # page_number = request.GET.get("page")
+        # page_obj = paginator.get_page(page_number)
 
-    context = {
-        "books": books,
-        "passages": passages,
-        # "paginator": paginator,
-        "book": book,
-        "passages_loaded": passages_loaded,
-    }
+        context.update(
+            {
+                "books": books,
+                "passages": passages,
+                # "paginator": paginator,
+                "book": book,
+                "passages_loaded": passages_loaded,
+            }
+        )
+        logger.info(f"Passages page used", request)
+    except Exception as e:
+        logger.error(f"Passages page error {e}", request, exc_info=True)
     return render(request, "passages.html", context)
 
 
-def passages_compare(request: HttpRequest) -> HttpResponse:
-    default_ids = [1]
-    passage_ids = request.GET.getlist("id", default_ids)
-    if not (2 <= len(passage_ids) <= 4):
-        passage_ids = default_ids
-    passages = passage_provider.get_passages_by_ids(passage_ids, as_json=True)
-    algorithms = algorithm_provider.get_configs(as_json=True)
-    context = {
-        "passages": passages,
-        "algorithms": algorithms,
-        "saved_algorithms": algorithms,
-        "algorithm_templates": algorithm_provider.get_default_configurations(),
-    }
+def passages_compare_page(request: HttpRequest) -> HttpResponse:
+    context = {}
+    try:
+        default_ids = [1]
+        passage_ids = request.GET.getlist("id", default_ids)
+        if not (2 <= len(passage_ids) <= 4):
+            passage_ids = default_ids
+        passages = passage_provider.get_passages_by_ids(passage_ids, as_json=True)
+        algorithms = algorithm_provider.get_configs(as_json=True)
+        context.update(
+            {
+                "passages": passages,
+                "algorithms": algorithms,
+                "saved_algorithms": algorithms,
+                "algorithm_templates": algorithm_provider.get_default_configurations(),
+            }
+        )
+        logger.info(f"Passage comparison page used with ids: {passage_ids}", request)
+    except Exception as e:
+        logger.error(f"Passage compare page error {e}", request, exc_info=True)
+
     return render(request, "passages_compare.html", context)
 
 
+# Deprecated
+# @csrf_exempt  # (deprecated?) TEMPORARY TODO
+# def algorithms2(request: HttpRequest) -> HttpResponse:
+#     VerbFormSet = formset_factory(VerbForm, extra=1)
+#     FrequencyFormSet = formset_factory(FrequencyForm, extra=1)
+#     verb_formset = VerbFormSet(prefix="verbs")
+#     frequency_formset = FrequencyFormSet(prefix="freqs")
+
+#     passages = passage_provider.get_all_passages(as_json=True)
+#     algorithm_templates = algorithm_provider.get_default_configurations()
+#     saved_algorithms = algorithm_provider.get_saved_algorithms(configs_only=True)
+#     context = {
+#         "algorithm_templates": algorithm_templates,
+#         "saved_algorithms": saved_algorithms,
+#         "verb_formset": verb_formset,
+#         "frequency_formset": frequency_formset,
+#         "passages": passages,
+#     }
+#     return render(request, "algorithms2.html", context)
+
+
 @csrf_exempt  # TEMPORARY TODO
-def algorithms2(request: HttpRequest) -> HttpResponse:
-    VerbFormSet = formset_factory(VerbForm, extra=1)
-    FrequencyFormSet = formset_factory(FrequencyForm, extra=1)
-    verb_formset = VerbFormSet(prefix="verbs")
-    frequency_formset = FrequencyFormSet(prefix="freqs")
-
-    passages = passage_provider.get_all_passages(as_json=True)
-    algorithm_templates = algorithm_provider.get_default_configurations()
-    saved_algorithms = algorithm_provider.get_saved_algorithms(configs_only=True)
-    context = {
-        "algorithm_templates": algorithm_templates,
-        "saved_algorithms": saved_algorithms,
-        "verb_formset": verb_formset,
-        "frequency_formset": frequency_formset,
-        "passages": passages,
-    }
-    return render(request, "algorithms2.html", context)
-
-
-@csrf_exempt  # TEMPORARY TODO
-def algorithms(request: HttpRequest) -> HttpResponse:
+def algorithms_page(request: HttpRequest) -> HttpResponse:
     context = {
         "algorithms": algorithm_provider.get_configs(as_json=True),
     }
+
+    logger.info(f"Algorithms page used", request)
     return render(request, "algorithms.html", context)
 
 
-def features(request: HttpRequest) -> HttpResponse:
+def features_page(request: HttpRequest) -> HttpResponse:
     features = bhsa_features_provider.get_features_for_display()
     context = {"features": features}
 
+    logger.info(f"Features page used", request)
     return render(request, "features.html", context)
 
 
-def settings(request: HttpRequest) -> HttpResponse:
+def settings_page(request: HttpRequest) -> HttpResponse:
     context = {}
 
+    logger.info(f"Settings page used", request)
     return render(request, "settings.html", context)
 
 
@@ -149,10 +179,15 @@ def settings(request: HttpRequest) -> HttpResponse:
 def get_hebrew_text(request: HttpRequest) -> JsonResponse:
     p_id = request.POST.get("passage_id")
     as_widget = request.POST.get("as_widget")
-    passage = passage_provider.get_passages_by_ids([p_id])[0]
+    passage: Passage = passage_provider.get_passages_by_ids([p_id])[0]
     words = word_provider.get_words_from_passage(passage, as_json=True)
     context = {"words": words}
     template = "hebrew_text"
+    logger.info(
+        f"Hebrew text widget fetched for {passage.get_reference()}",
+        request,
+        extra={"passage": passage.to_dict()},
+    )
     if as_widget:
         context.update({"passage": passage.to_dict()})
         template = "passage_text_widget"
@@ -160,7 +195,8 @@ def get_hebrew_text(request: HttpRequest) -> JsonResponse:
     return JsonResponse({"html": text_html})
 
 
-def run(request: HttpRequest, algorithm: Algorithm):
+# TODO : needs implimented again -- possibly
+def run_algorithm_from_config_form(request: HttpRequest, algorithm: Algorithm):
     data: dict = json.loads(request.body)
     config = algorithm.as_config(True)
     response = {
@@ -210,11 +246,16 @@ def algorithm_form(request: HttpRequest) -> JsonResponse:
                     algorithm = form.update_algorithm()
 
                     if action == "RUN":
-                        response = run(request, algorithm)
+                        response = run_algorithm_from_config_form(request, algorithm)
                         return JsonResponse(response)
 
                     else:
                         algorithm.save()
+                        logger.info(
+                            f"Algorithm saved/updated: {algorithm.name}",
+                            request,
+                            extra={"algorithm": algorithm.as_config(as_json=True)},
+                        )
                     return JsonResponse(
                         {
                             "algorithm": algorithm.as_config(True),
@@ -222,6 +263,11 @@ def algorithm_form(request: HttpRequest) -> JsonResponse:
                         }
                     )
                 else:
+                    logger.warning(
+                        f"Algorithm base form invalid",
+                        request,
+                        extra={"errors": form.get_errors()},
+                    )
                     return JsonResponse(
                         {
                             "status": "error",
@@ -232,7 +278,11 @@ def algorithm_form(request: HttpRequest) -> JsonResponse:
                     )
             except Exception as e:
                 exception = e
+                logger.error(f"Algorithm form exception {e}", request, exc_info=True)
         else:
+            logger.warning(
+                f"Algorithm form invalid", request, extra={"errors": form.get_errors()}
+            )
             return JsonResponse(
                 {
                     "status": "error",
@@ -268,6 +318,7 @@ def check_data_ready(request: HttpRequest) -> JsonResponse:
     return JsonResponse({"data_loaded": data_loaded})
 
 
+# Depricated? not referenced anywhere in the js currently.
 @require_POST
 @csrf_exempt  # TEMPORARY TODO
 def post_algorithm(request: HttpRequest) -> JsonResponse:
@@ -299,6 +350,7 @@ def post_algorithm(request: HttpRequest) -> JsonResponse:
     return JsonResponse(response)
 
 
+# Depricated? not referenced anywhere in the js currently.
 @require_POST
 @csrf_exempt  # TEMPORARY TODO
 def run_algorithm(request: HttpRequest) -> JsonResponse:
@@ -318,8 +370,10 @@ def run_algorithm(request: HttpRequest) -> JsonResponse:
             return JsonResponse({"status": "success", "data": response})
 
         except Exception as e:
+            logger.error(f"Failed to run algorithm {e}", request, exc_info=True)
             traceback.print_exc()
             return JsonResponse({"status": "error", "message": str(e)})
+    logger.warning(f"Invalid data to run the algorithm", request)
     return JsonResponse(
         {"status": "error", "message": "invalid data " + request.POST.dict()}
     )
@@ -341,6 +395,11 @@ def post_algorithm_comparisons(request: HttpRequest):
 
         passages = passage_provider.get_easiest_passages(count)
 
+        logger.info(
+            f"Comparing {count} passages with algorithms {alg1.name} and {alg2.name}",
+            request,
+        )
+
         for passage in passages:
             for i, config in enumerate([alg1, alg2]):
                 score, penalties = alg.get_passage_weight_x(config, passage)
@@ -357,6 +416,7 @@ def post_algorithm_comparisons(request: HttpRequest):
         )
 
     except Exception as e:
+        logger.error(f"Failed to compare algorithms {e}", request, exc_info=True)
         traceback.print_exc()
         return JsonResponse({"status": "error", "message": str(e)})
 
@@ -366,22 +426,51 @@ def post_algorithm_comparisons(request: HttpRequest):
     return JsonResponse({"status": "success", "html": comparisons_html})
 
 
+def verify_password(request: HttpRequest):
+    password = request.POST.get("password")
+    if password == settings.SECRET_PASSWORD:
+        logger.info("Deletion successful", request)
+        return True
+    return False
+
+
 @require_POST
 def delete_words(request: HttpRequest) -> HttpResponseRedirect:
-    word_provider.delete_all()
-    return redirect("settings")  # Redirect to the settings page
+    logger.warning("Attempting to delete all words", request)
+    success = verify_password(request)
+    if success:
+        word_provider.delete_all()
+    return render(
+        request,
+        "settings.html",
+        {"success": "Words deleted" if success else "Wrong password"},
+    )
 
 
 @require_POST
 def delete_passages(request: HttpRequest) -> HttpResponseRedirect:
-    passage_provider.delete_all_passages()
-    return redirect("settings")  # Redirect to the settings page
+    logger.warning("Attempting to delete all passages", request)
+    success = verify_password(request)
+    if success:
+        passage_provider.delete_all_passages()
+    return render(
+        request,
+        "settings.html",
+        {"success": "Passages deleted" if success else "Wrong password"},
+    )
 
 
 @require_POST
 def delete_algorithms(request: HttpRequest) -> HttpResponseRedirect:
-    Algorithm.objects.all().delete()
-    return redirect("settings")  # Redirect to the settings page
+    logger.warning("Attempting to delete all algorithms", request)
+    success = verify_password(request)
+    if success:
+        Algorithm.objects.all().delete()
+    return render(
+        request,
+        "settings.html",
+        {"success": "Algorithms deleted" if success else "Wrong password"},
+    )
 
 
 @require_POST
